refactor: rename internal python module
