+++ python/deltalake/__init__.py
@@ -1,5 +1,5 @@
+from ._internal import PyDeltaTableError, RawDeltaTable, rust_core_version
 from .data_catalog import DataCatalog
-from .deltalake import PyDeltaTableError, RawDeltaTable, rust_core_version
 from .schema import DataType, Field, Schema
 from .table import DeltaTable, Metadata
 from .writer import write_deltalake
