+++ python/deltalake/_internal.pyi
