+++ python/deltalake/schema.py
@@ -1,6 +1,6 @@
 from typing import Union
 
-from .deltalake import ArrayType, Field, MapType, PrimitiveType, Schema, StructType
+from ._internal import ArrayType, Field, MapType, PrimitiveType, Schema, StructType
 
 # Can't implement inheritance (see note in src/schema.rs), so this is next
 # best thing.
